% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichFIESTA.R
\name{whichFIESTA}
\alias{whichFIESTA}
\title{Locate a given FIESTA or FIESTAutils function}
\usage{
whichFIESTA(func_name)
}
\arguments{
\item{func_name}{A string representing the function name you wish to locate}
}
\description{
Locate a given FIESTA or FIESTAutils function
}
\examples{
whichFIESTA("pcheck.varchar")
}
